refactor/safe_authenticator: refactor app revocation logic
