+++ safe_authenticator/Cargo.toml
@@ -15,7 +15,8 @@ ffi_utils = { path = "../ffi_utils", version = "~0.1.0" }
 futures = "~0.1.14"
 log = "~0.3.7"
 maidsafe_utilities = "~0.13.0"
-routing = "~0.31.0"
+# routing = "~0.31.0"
+routing = { path = "../../routing" }
 rust_sodium = "~0.3.0"
 serde = "~1.0.9"
 serde_derive = "~1.0.9"
